requirements.txt
streamlit
langchain
huggingface_hub
python-dotenv
